pyproject.toml
[project]
name = "mcp-datos-abiertos-col"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.13.1",
    "selenium>=4.35.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
